Fix How The Dungeon Map is Accessed
currently, the dungeon map screen is displayed by pressing
the left shoulder button - which makes no sense whatsoever,
but is a work around to a Monogame + UWP back button
implementation bug / decision.

the dugeon map screen should be displayed not by a button press,
but instead by selecting the map menuItem from the inventory screen
by using the A button press, from the loadout widget.

when this series of events happens, the map is displayed.
it's kind of a pita because I don't want the map being used as a
crutch for players, plus this fixes the back button UWP issue.

updated Input.MapPlayerInput(), removing paths for shoulder
button input for inventory and map screens. only the start
button opens the map screen.

updated inventoryScreen.HandleInput() to create a new map
screen if the selected menuItem is the loadout's map.

playtested, working as expected.
this has additional benefit of removing a button from the total
input required to play the game. 'select' button is no longer needed.