EnemyTools Widget
created enemy tools widget, which for now simply
contains enemy spawn objects. in the near future,
we're going to improve this system so that the spawn
objects look like the enemy that they spawn.
easy to setup and install, working as expected.
we can spawn enemies in the editor again. kinda.