Implement Overworld Levels
the game currently only has 'rooms', which are building
interiors, essentially. we want outdoor levels, without
walls or exit objects, where the player feels more free
to roam. however, these outdoor levels still need to be
kept small, due to game design constraints.

so the idea is to implement these outdoor levels using
the same room recs and classes that interior rooms use.
when the hero stops colliding with a field level's room rec,
then the hero simply exits back to the overworld screen.
this has the effect of allowing the player to walk off
screen to exit any overworld / outdoor level. this approach
also makes outdoor levels easier to work with, because they
have exits by default, where as the dungeon's exit room
MUST have an exit roomObject to return to overworld screen.

to implement all this required changing a bunch of stuff,
over the course of several days, with lots of playtesting.

also, camera tracking to hero has been disabled for now.
the camera simply centers itself to the current room, and
quickly flys between rooms inside the dungeon, which is
similar to how the original zelda handled the camera.