Organization++
moved AddDevDoors() into Dungeon Functions, because thats where it belongs.
reorganized BuildRoom() a bit.

created ShutDoors(), which is responsible for shutting any and all doors
in a room, which by shut i mean convert them to 1 way trap doors which
push any actors or objects in their facing direction with max force.

updated ProcedurallyFinish() to ShutDoors() on the boss room and the
dev boss room, while also not spawning the boss in the dev boss room,
so it's easier to develop and test.

playtested, it's all working quite well actually...