Dungeon Secret Rooms ++
tripled the amount of secret rooms added to dungeons. because
they are awesome. and i kinda want to be able to navigate a dungeon
entirely by secret rooms. playtesting confirms that more secret
rooms are awesome, which led to an idea for the dungeon 'recipe'.

(at the end of the dungeon algorithm, it adds additional rooms)
altered the dungeon recipe to add 1 wave of additional rooms, THEN
add two waves of secret rooms, then do it again. this has the net
effect of spreading the dungeon out a bit more, while also connecting
it more with secret rooms. this allows the player to explore the hidden
side of the dungeon thru secret rooms, or take a more straight forward
approach by traversing thru doors - since those critical paths are built
first. it's a pretty amazing recipe and i'm super happy with it.

updated AddSecretRooms() to not attach to HUB room, so as to prevent
the case where a secret room could connect hub and key rooms, making
the dungeon's critical path much too short to be challenging.