Game Recordings ++
fixed a variety of chest related issues with hub and key rooms,
while recording with obs studio, which i was able to pickup and learn
in about 2 minutes because it's interface isn't hot garbage - thanks!
now i'm able to record 60fps .mov from hardware and they look criisp.
so, i'll post a playthrough of where the game is right now, along
with a few select gifs coming up real soooon....