Loot++
loot should spawn centered to the obj that was destroyed.
the issue was simply that pickups = loot, and pickups are
8x16 pixels, where most objs that spawn loot are 16x16, and
of course i wasn't adding in an x axis offset, so all loot
spawned biased to the left. added an offset in for the x axis,
playtested, looking great now.