DungeonRooms++
we should be storing if a puzzle room has been solved or not.
we may need to add a boolean here to room, in order to track if hero has
solved it. if so, then upon a room build, the trap doors should be opened.
this adds a sense of progression to the dungeon as well.

*smart* if we build a puzzle room thats been visited before, DONT setup
any puzzles in the room. the player has already solved this puzzle room.
implemented, playtested, working as expected.