Items++
moved net to ITEMS X category in inventory screen.
this is actually where the net belongs, because it's modeled
as a weapon internally. why it was placed in the items category
is simply my mistake.

changed quest items to MISC, because i have no idea what
this is actually going to be named and quest items sounds corny af rn.
playtested, working.