Implement Pickup + Carry
hero, and eventually all actors, will be able to pickup
and carry skull pots, world pots, and small bushes, just
like LttP. i already implemented this in a previous
version, but that implementation broke so i ripped or
commented out all that code. looking at it with a fresh
perspective, pickup carrying and eventually throwing
objects is something that all actors should be able to
do, not just the hero. with that in mind, actor methods
were created to handle picking up and carrying objs.

a temporary throw() method was created, but it just
destroys the held obj. this feeds into the handle common()
method, which then controls creating the proper explosion
of leaves or debris for the destroyed object.

playtested, working quite well.

falling into pits with held obj: x
getting hit with held obj: x
exiting room with held obj: well, actually...

the way the system works now, players can pickup one obj
from RoomA and carry it to RoomB. theoretically, they could
carry an obj through as many rooms as they want to, since
the object is forced to remain active. this means it *should*
live through Pool.Reset() when we change rooms, and it def
wont be returned from Pool.GetRoomObj() because it's active.
so it should be protected/safe to carry from room to room.

playtested exiting dungeons / levels while carrying objs,
the obj dissappears upon a level build, which is great.
transitioning between rooms, the heldObj is untouched.