DungeonRooms++
created Row006, which combines switches with blocks with switchBlocks.
my naming conventions are failing me. but the rooms fun.

created Column006, which teaches player they can use dungeon
statues (or barrels) to activate the switch in the middle of the room.
what's cool about this room is that the conveyor belts both
point to the switch, and will move the dungeon statues onto
the switch, opening the doors - so they're super useful.

created Row007, which resembles an engine - moving parts
in a stuttered rhythm going around blocks that alternate
up and down, like cylinder heads. this is a switch with
moveable block puzzle, just decorated to be distracting.

created Column007, which requires the player to move
one of two light blocks to the switch in the center
of the room. however, this is done on 'platforms',
which are areas of walkable ground surrounded by pits.
so the hero can fall in, the blocks can fall in, etc..
*this room can be put into a state that is unbeatable,
if hero pushes all the light blocks into the pits.
later on, when we have the cane, we could create a block
on the switch to solve the puzzle.. and we'd have
the magic mirror to warp us to the dungeon exit.
so it's fine.

fixed a minor problem with Square002.
improved the visuals on Square009.
recompiled roomData. playtested game. fun as hell.
oh btw, you can push blocks diagonally. i found that
out by trying it, and it just works.