Enemies & Pits ++
enemies should not fall into pits unless they are in a hit state.

test that enemies cant walk into pits. passed.
test that striking enemies pushes them into pits. oh yeah.
test that hero can easily walk into pits. yup.
test that hero isn't pulled into pitTraps if he's on the edge. eh..
there were some complications here..

we were setting the pit's lifetime to 100 as a secret
signal to the following code that the pit was created
this frame by an interaction with the pit trap obj.
this was clever clever, but not necessary, as the
better solution is simply to return from the method
after hero interacts with pit trap, instead of continuing
to evaluate the proceeding code, which incorrectly
represents the pit obj with the pitTrap obj's hitBox.
this causes the known issue where the hero can't escape
being pulled into a pit trap, which feels cheap.

lots of playtesting.

if link dies in the dungeon, he shows up on the overworld
screen still in his sitting / dead animation. but as soon
as the player presses a direction or A, link stands back
up and everything's back to business. awesome.
that happens because we're using Pool.hero on overworld
screen, instead of creating an instance like we were
before. it's simply the last state the hero was in.