Rooms++
issue with key room: the ice conveyor belt room.
the wall statues can shoot arrows that trigger the lever,
rendering the room super easy to beat. fixed.
various room improvements and tweaks.
some code cleanup.

Editor++

fixed a super annoying issue with editor: we set hero
to have everything while in dev mode, meaning he has the
dungeon map, while in editor mode. when we load a room,
the room building routine checks to see if hero has the
map, and if he does the room builds an empty chest instead
of a key chest (assuming it's living in a dungeon, when
it isn't). this means the key room loads into editor with
an empty chest, instead of a key chest. it's easy to not
notice it, and then save the roomData xml, which means in
game that room has an empty chest and the player can't get
the key, which means they can't beat the boss / dungeon.
fixing this was super easy: when room editor loads up,
we just set Level.map to false. that's it.