Swim Interactions
updated hero interactions to allow for most interactions
while in the water, including: activating torches, levers,
switch block buttons, chests, and the boss door. so, we
*could* fill the hub room with water tiles and have a miniboss
battle on platforms - if that seems fun. lol.

the reason this was done was so that hero can interact
with objs in dungeon rooms while swimming, so water tiles
can become a part of dungeon rooms, expanding the room variety.