Improvements++
add burnt tree to world widget
updated forest entrance + colliseum levels.
improved sorting of water tiles.
added two dungeon rooms with water tiles.
improved blob death animation to be blank.

updated interaction routines to only play water walking
sound if the hero is walking on coastlines. it was
super annoying to hear that sound spammed when the room
was full of enemies. though accurate, not desired.

fixed a bug where pressing A while in the water caused hero
to interact with the water tile, completely stopping his
inertia, due to a StopMovement() call. now we're just
ignoring the water obj in CheckInteractionRec().

renamed CheckInteractionRecCollisions() to CheckInteractionRec().
adding 'collisions' to the end of that makes no sense.
lots of playtesting in editor + game modes.

yass: so... you can pickup some objects. and you
can jump in the water while carrying this obj, and
it will stay above your head, and you can't throw it. yet.
and you can swim, exit the water, and throw the obj.
glorious. i didn't even consider that state of the
game until i tested it, and it passed.

fixed a bug where hero couldn't die in game.
upon death, his health was reset to 3. oops.
that fixed one problem, but there are more.
i sleep now. more tmrw.