Hero Death++
fixed a major bug where hero couldn't die in game.
this was due to SetDeathState() setting Actor.stateLocked true,
then calling Transform(), which called SetType(), which called
ResetActor(), which set statelocked to false. this caused the hero
to be able to accept input while in a death state, which
led to the buggy mess of being alive and dead at the same time.