Improvements++
improved row rooms that were looking bland.
improved editor to reset level key and map booleans
to false just prior to building a room from xml file.
this ensures that the room will be built properly,
even if we have key and map cheats enabled. those cheats
are checked when a level  or dungeon is built,
not when a room is built.

improved hub rooms. now they are setup to handle a
miniboss, which will come very soon.
recompiled roomData.