Shovel MenuItem + Data
created shovel sprite, to match the diagonal orientation
of the other items and weapons used in the game. my theory
on why the shovel was vertical and not diagonal in LttP
is likely because it was a transitory item, meant to be
used for a short period of time, replaced by the flute,
after link digs in the top left area in the circle of trees.
so, either the visual designers were cue-ing the player
that this item is different (disposabe), or they were
just being lazy or re-using an old sprite.

added MenuItemType WeaponShovel.
added animFrame and menuItem.type() def.

added saveData.weaponShovel boolean.
updated unlock all cheat to unlock shovel.

updated inventory widget to display shovel, next to net.
updated ui item sprite sheet.
playtested, shovel menuItem is in game and looking good.