SaveData++
we need to convert weapons, armor, and equipment to using the system
that items uses, because it scales much better and doesn't abstract
the item to an int value, but rather an enum string representation,
so it's easier to save, load, and work with. however, this means
savedata prior to this commit wont load. *soft violins play*

fixed how weapons, armor and equipment were stored in savedata.
removed all equipment booleans other than ring from the game
(since they weren't being implemented anyway).

updated equip routines to work with new system.

removed Fun_Debug.Inspect(SaveData). we can just save the
game and open the xml data. not needed.

playtested, working very well:
the game handled the corrupt savedata fine, overwriting it
with the current savedata in memory so the game doesn't crash,
and the player can still save their progress. current savedata
in memory is created fresh when the game loads, so it's just
a new game essentially. this means the old savedata file is
lost forever tho, which sucks, but also is kinda to be expected
since this game is in development and free.