Dig()
created Functions_Hero.Dig(), called from Actor functions Update(),
when actor is on land, non-carrying, using an item, and the item
equipped currently is the shovel. i've essentially 'layered' this
into the UseItem() system currently in place - the shovel projectile
programmed previously will still function as before, destroying objects
and hurting enemies - but now we have a call to Dig() just before it.

implemented Dig() for testing purposes, it currently
creates and places a tall grass object aligned to the game
grid, so it's 80% working. wow. we just need to bolt on a
few checks and add some ditches. very close now.