Ditch System++
created sprites for ditch system, with empty and filled
versions, so we can fill ditches up with water.

created ObjTypes for empty and filled ditches (16 new objs).
created obj.Defs + animFrames for empty and filled ditches.
upon writing the ditches system, i realized i needed
more ditch sprites: endcaps N, S, and horizontal, plus
a 3 connector that was vertical.

add those objTypes, animFrames, and typeDefs.
completed working basic version of ditch system,
where two ditches can be connected as endcaps.
added in corners, 3 connections, and 3 connections.
lots of playtesting here. lots.

created Functions_Dig to contain the increasingly
complex code managing ditches. happy to say it's
working very well.

added a final step to digging routine that updates
the surrounding ditches, so the ditch system
dynamically adapts itself to whatever tile the
player Dug() last. which is AWESOME. you know what
else is awesome? if hero digs on a tile with an obj
that shovel can destroy, then adding the ditch doesn't
increase the roomObj count - because we released the
destroyed object. so adding ditches doesn't incur
that much roomObj overhead.

added functions_dig to uwp project. built and
playtested uwp project. shoveling working well
on all platforms. it's a little expensive for my
tastes (the dig algorithm in total), but it gets
the job done right and the rest of the systems
are extremely low cost-wise, so it's ok.

digging is fun af and i did it for like 15 mins
after the system was completed, testing it.
it's beautiful.