Ditch Sprites++
on the edges of the ditches, we should be using transparent black
pixels. currently we're using 100% opaque pixels, and they're clashing
with the water tiles. if we switch to using black with an alpha, then
they'll blend nicely with water tiles and look good.

once these are looking good, migrate those ground
changes to the filled ditch sprites.

- so this was done, and the idea behind it is solid,
however, water tiles are a darker shade than the grass color,
which is the default background color. when we add a
ditch obj, we remove the coastline obj, removing the
darker background revealing the lighter background,
and the semi-transparent pixels 'blend' over the bkg
color just fine - but it's the grass color, not the
surrounding water tiles color. the net effect being,
there is no visual change between previous sprites
and new sprites. lol. had to try.