Filled Ditches ++
if hero interacts with a ditch that hasAI, then that ditch
is filled, and this obj should be treated the same way we treat
coastlines, so it can be included in the coastlines check..

during the process of implementing this, the routine
necessary to show the actor's feet in the animated ripple
state was duplicated, leading to the creation of
Functions_Actor.DisplayWetFeet(). this is called from
an actor's interaction with coastlines (current sys),
and now it's being called from actor interactions
with filled ditches - since they display water.
tested, working great.