Improve Filled Ditch Visuals
we also need a better way of handling how filled ditches visually
transition into water tiles. the system works fine, but the visuals
could be improved. essentially, the ditch touching the water tile
needs to have the water tile's sprite undeath it, so it's bkg matches
the bkg of the water tile. we *could* create a floor sprite at ditches
position, then set the floor tile's texture and animFrame (to water tiles).
this is likely the best approach to getting ditches looking 100%.

improved Dig() to get a floor tile and place it at the ditches
position, if the ditch touches a water tile. this means the ditch
is a root ditch for spreading water to other empty ditches. by
adding the water background, the root ditch blends much better
with the water tile. playtesting this outside went great.
when I took it to the dungeon for break-testing, I found
a ton of bugs, some hilarious, some game-breaking.
So I feex.