Dig Testing
ok, so there are some edge cases here that I really didn't think about.
for example, link can dig out switch blocks that are down, because they
aren't blocking. along with conveyor belts. and switches.
probably spike blocks. floor spikes. bumpers. pets? yikes.
and icetiles... wait a second.. no..

Digging and Ice Tiles
we allow link to dig out ice tiles, because this seems plausible in the world,
and those ice tiles can be a real pain. now, we have a logical way of
'defeating' them - which I absolutely LOVE and I'm keeping it.

..continuing on, updated Dig()'s obj type check with
section for switch blocks, conveyor belts, switches,
spike blocks, spipke floors, bumpers, fairies, and pets.

added a commented out section of code explaining that
ice tiles are allowed to be dug out. in the future,
this may change if the game is put into hard mode.
however, this may be a bad idea - because I don't
want the player's interaction with the world to change
in hard mode, only the 'systems' that have been built
to help them are (mostly) turned off.