Exits++
Created exit roomdata 1 and 2, containing very basic
and easy switch puzzles. Its an easy early victory.
Updated level systems to handle exit roomdata loading.

Added skull pillar object
this obj exists inside of a dungeon's exit room, to spruce
it up and give the player a sense that this is a different
place than the outside they just left. also, exit rooms
contain simple switch puzzles for the player to complete,
so skull pillars were updated to be pushable / moveable,
and then integrated into the existing exit room puzzels.
since this room is the only place where these skull pillars
exist in game now, this adds some salt to the otherwise
pretty bland exit room.

various comment cleanups too.

changed how hero's grab() check looks for a grab-able
object: we were doing type checks, which made sense,
but now we're doing obj.compMove.moveable checks,
which make even more sense - if the player sees an
object that can be moved by a conveyor belt, then
it makes sense that the object is moveable, and the
player should be able to move it by pushing it.
the issue is that i'm not sure exactly which objs
are moveable, and if i want all those objs pushable
too. this also reduces the complexity of grab routine.

updated tables + shelves to be moveable.
playtested, woroking as expected.