The Farm
create The Farm level.
this is Link's new house, as an adult.
for now, it's just open property.
in the future, it will be locked off
until link completes a questline to
open it up. this is entirely optional
to the main story line.

Why?
1. primary reason is to give player a level
to call home. eventually there should be a
carpenter NPC that allows the player to buy
new blueprints for different building styles.
how big the building is affects how much land
crops can be grown on, which affects how much
profit and loot the player can harvest at once.
2. to introduce the idea of farming, which uses
ditches + water tiles to irrigate crops (bushes).
all we need to do is make a bush stump obj
become a bush obj - if a filled ditch is a
neighboring tile. this allows for "farming" of
bushes for loot - quite literally. perfect.

The Setup / Puzzle - The Dead Farm
initially, each time hero enters the farm level,
there is a series of ditches dug around some bushes,
near a body of water - but there is no connecting
ditch, and thus the ditches are dry.
the bushes are actually bush stumps (upgraded
with grow routines, via obj.GetsAi = true).
the puzzle is "solved" by digging a ditch
from the pond to the irrigation ditches,
filling them with water. once this happens,
the bush stumps grow into bushes.
when link cuts a bush, loot may drop,
but the bush will always create a stump,
which will then be evaluated to grow
again - leading to a state where player
can farm the bushes for loot. we want the
player to have loot. we want loot to be
easy to get. grinding doesn't have to be
hard - it can be fun.
Bonus: leetle tribute to Stardew Valley.

created farm level.
changed Church levelID value to TheFarm.
changed Church roomID to TheFarm.
updated RoomData.
Church level is now Farm level.
changed active location on overworld map.
created a shortcut on map to get to
TheFarm, for quick testing.

Design++
design the farm level better based on the map sprite.
..and based on this design, we need ALOT more roomObjs.
this maybe the level that tests the limits of the pool.
which is good. let's get this out of the way NOW.

Pool++
doubles roomObj max from 750 to 1500.
let's see how my systems scale! ... no real change.
increased to 3000. and we'll need to check this on the
xbox one too, due to decreased cpu power,
large levels like the farm may test the 60fps rule.

improved farm level with coastlines and a
small farm of 80 bushes, with irrigation
ditches already dug. farm roomObj total
is now at 1219, a factor of 10 larger than
previous levels. so we're testing what the
systems are capable of now. on the pc,
we're STILL at around 1ms.