Editor++
when we press F3, the top menu is hidden, but we can
still click on the buttons, which is bad when we're
trying to complete a level. fixed. input now returns
from the handleInput method post F1-5 key evaluation.
modified update to still allow for mouse input.
that is a super useful editor addition.