Turtles++
turtles can now be picked up, carried,
and thrown - and the turtle doesn't get
hurt during this process.

Bug: held turtles + water
if we carry turtles into water, they want
to be released, and create splash particles.
so we need to check and see if obj.compCollision
is false, and if so - dont create splashes or
release the obj. done. link can now pickup
and carry turtles into the water with him,
and hide underwater with a turtle over his
head, allowing him to sneak as a turtle.
turtle power.