Enemies & Levels++
added rats to dungeon rooms, and forest entrance level.
improved farm level with beyond level edge water tiles,
so hero doesn't splash out of water walking off screen.

Miniboss++

fixed a bug where miniboss didnt have swim animFrames,
leading to a null ref exception. this bug is likely happening
because the miniboss spawns in, then is evaluated to be in a
swimming state. he shouldn't be evaluating to be in a swimming
state, because he's flying.

changed miniboss to shoot arrows, because shooting
fireballs was too powerful (for a miniboss).