I might be Dumb - Part 1
So, I added a return; to the end of InteractActor()
in the hero vs door routine, which bailed from the method,
preventing trap doors from ever evaluating against
the hero - which allowed hero to freely walk thru trap
doors, completely breaking rooms + puzzles + dungeons.
and since I'm a AAA developer, I haven't noticed this
until now, and I don't remember when I put the return;
statement in there - so the game may of been in this broken
state for a while. but that's not from laziness, it's
simply because there is so much to do and fix.

anyway, I removed the return and trap doors work again.
sorry about that.