Lttp AND Hld Dashing
in hyper light drifter, pressing the dash button
quickly moves the hero in the input direction for a
brief period of time. it's tactical.

in lttp, pressing and holding the dash button
"winds up" link to do a long-term dash in a single
direction. not tactical, more used to traverse
large areas quickly.

both of these dashing approaches have their own
pros and cons. currently, the game implements hld
style dashing, but if change the input slightly,
then we can get a unique version of lttp's dashing too.

Dash Upgrade
if player holds down the B button while pressing
an input direction, then link will dash in that
direction. this isn't a new button press - we're
now checking to see if the B button is down, which
essentially spams dash, while still allowing for
directional control. this is an instant tactical
dash combined with distance dashing - plus player
can change direction while dashing. I see this new
dashing system as superior to lttp & hld, and almost
feels overpowered. so we'll scale the challenges up
to meet the new powerful skill introduced - instead
of nerfing the skill and killing the fun.