Bombos
updated ui items sheet with bombos magic sprite.
created magicBombos menuItem value.
added magicBombos saveData boolean.
implemented a temporary version of bombos
that creates a line of explosions in the
caster's direction - which isnt exactly
what I want, but it will work for now.

Magic++
improved description for fireball and bombos
magic menuItems.

Thoughts
so, we implemented a series of destructive
projectiles that exist for only a moment (explosions).
if we replace the sprite with a lightning bolt,
and the sound fx with lighting, and implement
the lighting menuItem, then plug it into vendor,
then we've just added BOLT magic to the game.
which is nice.